% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_is_spline.R, R/meth_is_spline_0ld.R
\name{is.splinets,Splinets-method}
\alias{is.splinets,Splinets-method}
\title{Diagnostics of splines}
\usage{
\S4method{is.splinets}{Splinets}(object)

\S4method{is.splinets}{Splinets}(object)
}
\arguments{
\item{object}{\code{Splinets} object, the object to be diagnosed;}
}
\value{
A list made of: a logical value \code{is}, a \code{Splinets} object \code{robject}, a numeric value \code{Er}, and a string \code{Report}..
\itemize{
\item The logical value \code{is} indicates if all the condtions for the elements of \code{Splinets} object to be a collection of valid
splines are satisfied, additional diagnostic messages are printed out.
\item The object \code{robject} is a modified input object that has all SLOT fields modified so the conditions/restrictions
to be a proper spline are satisfied.
\item The numeric value \code{Er} is giving the total squared error of deviation of the input matrix of derivative from the conditions required for a spline. 
\item The string \code{Report} contains a detailed report on the diagnostic, use \code{cat(is.splinet$Report)} to print out the report. 
}

A list made of: a logical value \code{is}, a \code{Splinets} object \code{robject}, and a numeric value \code{Er}.
\itemize{
\item The logical value \code{is} indicates if all the condtions for the elements of \code{Splinets} object to be a collection of valid
splines are satisfied, additional diagnostic messages are printed out.
\item The object \code{robject} is a modified input object that has all SLOT fields modified so the conditions/restrictions
to be a proper spline are satisfied.
\item The numeric value \code{Er} is giving the total squared error of deviation of the input matrix of derivative from the conditions required for a spline. 
}
}
\description{
This short information is added to satisfy an R-package building requirement, see \code{\link{is.splinets}} 
for the actual information.

This short information is added to satisfy an R-package building requirement, see \code{\link{is.splinets}} 
for the actual information.
}
