% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_plot.R
\name{plot,Splinets-method}
\alias{plot,Splinets-method}
\title{Plotting splines}
\usage{
\S4method{plot}{Splinets}(
  object,
  x = NULL,
  sID = NULL,
  vknots = TRUE,
  type = "stnd",
  mrgn = 2,
  lwd = 2,
  ...
)
}
\arguments{
\item{object}{\code{Splinets} object;}

\item{x}{vector, specifying where the splines will be evaluated for the plots.}

\item{sID}{vector, specifying indices of splines in the splinet object to be plotted;}

\item{vknots}{logic, indicates if add auxiliary lines to highlight the positions of knots; The default is \code{TRUE}.}

\item{type}{string, controls the layout of graphs; The following options are available
#' \itemize{
  \item \code{"stnd"} \code{object@type="dspnt"} or \code{="spnt"}, then it is ploted over 
  the dyadic net of supports, other bases are on a single plot with information about 
  the basis,
  \item \code{"simple"} all the objects are ploted in a single plot with minimal information,
  \item \code{"dyadic"} if not \code{object@type="sp"}, then the plot is over the dyadic net of supports.
}}

\item{mrgn}{number, specyfying the margin size in the dydadic structure plot;}

\item{lwd}{positive integer, the line width;}

\item{...}{also other graphical parameters can be passed;}
}
\value{
A plot visualizing a Splinet object. The entire set of splines will 
be displayed in a plot.
}
\description{
The method provides graphical visualization of a \code{Splinets}-class object.
}
\details{
Standard method of plotting splines in a \code{Splinet}-object. 
It plots a single graph with all splines in the object except if \code{type} field of the
object represents a splinet. In the latter case, the default is the (dyadic) net plot of 
the basis. The string argument \code{type} can overide this to produce a single plot.
Most of the standard R graphical parameters can be passed by this function.
}
\section{References}{

Liu, X., Nassar, H., Podg\eqn{\mbox{\'o}}{o}rski, K. (2019) Splinets -- efficient orthonormalization of the B-splines.  ArXiv, abs/1910.07341.

???? (2020) \code{Splinets} -- Local Taylor expansion driven representation of splines, their support sets and orthogonal bases. (Preprint)
}

\examples{
#-----------------------------------------------------#
#-------------------Ploting splinets------------------#
#-----------------------------------------------------#
#Constructed splines

n=25; xi=sort(runif(n+2)); xi[1]=0; xi[n+2]=1; k=3
supp=list(t(c(2,12)),t(c(4,20)),t(c(6,25))) #defining support ranges for three splines

#Initial random matrices of the derivative for each spline
SS1=matrix(rnorm((supp[[1]][1,2]-supp[[1]][1,1]+1)*(k+1)),ncol=(k+1)) 
SS2=matrix(rnorm((supp[[2]][1,2]-supp[[2]][1,1]+1)*(k+1)),ncol=(k+1)) 
SS3=matrix(rnorm((supp[[3]][1,2]-supp[[3]][1,1]+1)*(k+1)),ncol=(k+1)) 

spl=construct(xi,k,SS1,supp[[1]]) #constructing the first correct spline
nspl=construct(xi,k,SS2,supp[[2]],'CRFC')
spl=gather(spl,nspl) #the second and the first ones
nspl=construct(xi,k,SS3,supp[[3]],'CRLC')
spl=gather(spl,nspl) #the third is added

plot(spl)
plot(spl,sID=c(1,3))
plot(spl,sID=2)
t = seq(0,0.5,length.out = 1000)
plot(spl, t, sID = 1)

#Random splines
n=17; k=4; xi=sort(runif(n+2)); xi[1]=0; xi[n+2]=1 
S=matrix(rnorm((n+2)*(k+1)),ncol=(k+1))
spl=construct(xi,k,S) 
plot(spl,main="Mean Spline",lty=2,lwd=2,xlab='')

RS=rspline(spl,5)
plot(RS,main="Random splines around the mean spline",ylim=3*range(spl@der[[1]][,1]) )
lines(spl,col='red',lwd=4,lty=2)

}
\seealso{
\code{\link{evspline}} for manually evaluating splines in a \code{Splinet}-object;
\code{\link{Splinets-class}} for the definition of the \code{Splinet}-class;
\code{\link{lines,Splinets-method}} for adding graphs to exisiting plots;
}
